refactor(Components): refactor compoenents tree
